docker/web/requirements.txt
selenium
mysql-connector-python
